+++ android/measure/src/main/java/sh/measure/android/events/SignalProcessor.kt
@@ -416,6 +416,9 @@ internal class SignalProcessorImpl(
         isSampled: Boolean,
     ): Boolean = when {
         configProvider.enableFullCollectionMode -> true
+        eventType == EventType.HTTP -> {
+            sampler.shouldSampleHttpEvent()
+        }
         eventType in DefaultConfig.JOURNEY_EVENTS -> {
             sampler.shouldTrackJourneyForSession(sessionId)
         }